Fix issue 23514 - Incorrect compilation when adding a 64-bit constant to a link-time address (#14657)
* Fix issue 23514 - Incorrect compilation when adding a 64-bit constant to a link-time address

The current cast to int causes a loss of precision, which can result in
incorrect compilation.

* Add test23514.d

* Enable optimization in test

* Improve test case

* Use ulong instead of size_t

* Restrict test to 64-bit targets

* Re-enable test for 32-bit targets

* Disable test for win64 and add comment

+++ compiler/src/dmd/e2ir.d
@@ -514,7 +514,7 @@ elem* toElem(Expression e, IRState *irs)
     {
         elem *e;
         Type tb = (se.op == EXP.symbolOffset) ? se.var.type.toBasetype() : se.type.toBasetype();
-        int offset = (se.op == EXP.symbolOffset) ? cast(int)(cast(SymOffExp)se).offset : 0;
+        long offset = (se.op == EXP.symbolOffset) ? cast(long)(cast(SymOffExp)se).offset : 0;
         VarDeclaration v = se.var.isVarDeclaration();
 
         //printf("[%s] SymbolExp.toElem('%s') %p, %s\n", se.loc.toChars(), se.toChars(), se, se.type.toChars());
